Update test config to @angular/cli
The change from angular-cli to @angular/cli introduced several config
changes. Angular cli config is now in .angular-cli.json. I
also removed ng-hal from the repo. It was causing a lot of
errors. The owner is making updates on the repo to make it
compatible with the newer verions of angular, but the
updates are not published yet.